+++ server/shared/db/model.go
@@ -3,5 +3,5 @@ package db
 import "database/sql"
 
 type Model interface {
-	ConnectDB(username string, password string, host string, port uint32, dbname string) (*sql.DB, any)
+	ConnectDB(username string, password string, host string, port uint16, dbname string) (*sql.DB, any)
 }
